chore: cleanup

+++ src/account.rs
@@ -1,11 +1,10 @@
-use serde::{Deserialize, Serialize};
-use std::collections::HashMap;
+use serde::Serialize;
 
 #[derive(Debug, Serialize)]
 pub struct Account {
     pub client: u16,
-    pub available: f64,
-    pub held: f64,
-    pub total: f64,
+    pub available: f32,
+    pub held: f32,
+    pub total: f32,
     pub locked: bool,
 }
